-updated project tracker actions
